Showing of course creator picture in AllCourses and MyCourses views.

+++ E-LearningSystem/E-LearningSystem.Web/Controllers/CoursesController.cs
@@ -7,26 +7,31 @@
     using E_LearningSystem.Data.Models;
     using E_LearningSystem.Infrastructure.Extensions;
     using Microsoft.AspNetCore.Authorization;
+    using E_LearningSystem.Services.Services.Courses.Models;
+    using AspNetCoreHero.ToastNotification.Abstractions;
 
     using static E_LearningSystem.Infrastructure.Constants.IdentityConstants;
-    using E_LearningSystem.Services.Services.Courses.Models;
 
     public class CoursesController : Controller
     {
         private readonly ICourseService courseService;
         private readonly IShoppingCartService shoppingCartService;
         private readonly ITrainerService trainerService;
+        private readonly INotyfService notyfService;
         private readonly UserManager<User> userManagerService;
 
 
         public CoursesController(
             ICourseService courseService,
             IShoppingCartService shoppingCartService,
+            INotyfService notyfService,
             ITrainerService trainerService,
+
             UserManager<User> userManager)
         {
             this.courseService = courseService;
             this.shoppingCartService = shoppingCartService;
+            this.notyfService = notyfService;
             this.trainerService = trainerService;
             this.userManagerService = userManager;
         }
@@ -46,7 +51,7 @@
         [HttpPost]
         [Authorize(Roles = $"{AdminRole}, {TrainerRole}")]
         public async Task<IActionResult> CreateCourse(CourseFormModel courseModel, IFormFile pictureFile)
-        {                 
+        {
             ModelState.Remove("Categories");
 
             if (!this.courseService.CheckIfCourseCategoryExists(courseModel.CategoryId))
@@ -77,10 +82,18 @@
         }
 
 
-        [HttpGet]      
+        [HttpGet]
         [Authorize(Roles = $"{AdminRole}, {TrainerRole}")]
         public async Task<IActionResult> EditCourse(int id, IFormFile pictureFile)
         {
+            int courseCreatorId = await this.courseService.GetCourseCreatorId(id);
+            int currentTrainerId = await this.trainerService.GetTrainerIdByUserId(User.Id());
+
+            if (currentTrainerId != courseCreatorId)
+            {
+                return Unauthorized();
+            }
+
             var course = await this.courseService.GetCourseById(id);
 
             if (course == null)
@@ -93,16 +106,25 @@
                 Name = course.Name,
                 Description = course.Description,
                 CategoryId = course.CategoryId,
+                Categories = await courseService.GetAllCourseCategories()
             };
+
+            bool isEdited = await this.courseService.EditCourse(
+                id,
+                courseFormModel.Name,
+                courseFormModel.Description,
+                courseFormModel.CategoryId,
+                pictureFile);
 
             return View(courseFormModel);
         }
 
 
-        [HttpPost]       
+        [HttpPost]
         [Authorize(Roles = $"{AdminRole}, {TrainerRole}")]
         public async Task<IActionResult> EditCourse(int id, CourseFormModel courseModel, IFormFile pictureFile)
         {
+
             if (!this.courseService.CheckIfCourseCategoryExists(courseModel.CategoryId))
             {
                 this.ModelState.AddModelError(nameof(courseModel.CategoryId), "Category does not exist.");
@@ -128,21 +150,38 @@
 
 
         public async Task<IActionResult> AllCourses()
-        {         
+        {
+            List<AllCoursesViewModel> allCoursesViewModel = new List<AllCoursesViewModel>();
             var courses = await courseService.GetAllCourses();
-       
-            var allCoursesViewModel = new AllCoursesViewModel
-            {           
-                AllCoursesServiceModel = courses
-            };
 
+            foreach (var course in courses)
+            {
+                var trainer = await this.trainerService.GetTrainerByTrainerId(course.TrainerId);
+
+                var viewModel = new AllCoursesViewModel
+                {
+                    Id = course.Id,
+                    Name = course.Name,
+                    Description = course.Description,   
+                    Price = course.Price,
+                    ImageUrl = course.ImageUrl,
+                    CategoryName = course.CategoryName,
+                    AssignedStudents = course.AssignedStudents,
+                    ProfileImageUrl = trainer.ProfileImageUrl,
+                    TrainerName = trainer.FullName
+                };
+
+                allCoursesViewModel.Add(viewModel);
+            }
+        
             return View(allCoursesViewModel);
         }
 
 
         [Authorize]
         public async Task<IActionResult> MyCourses()
-        {           
+        {
+            List<AllCoursesViewModel> myCoursesViewModel = new List<AllCoursesViewModel>();
             IEnumerable<AllCoursesServiceModel> myCourses = null;
 
             if (User.IsInRole(TrainerRole))
@@ -154,11 +193,26 @@
             {
                 myCourses = await courseService.GetMyCourses(User.Id());
             }
-      
-            var myCoursesViewModel = new AllCoursesViewModel
-            {              
-                AllCoursesServiceModel = myCourses
-            };
+
+            foreach (var course in myCourses)
+            {
+                var trainer = await this.trainerService.GetTrainerByTrainerId(course.TrainerId);
+
+                var viewModel = new AllCoursesViewModel
+                {
+                    Id = course.Id,
+                    Name = course.Name,
+                    Description = course.Description,
+                    Price = course.Price,
+                    ImageUrl = course.ImageUrl,
+                    CategoryName = course.CategoryName,
+                    AssignedStudents = course.AssignedStudents,
+                    ProfileImageUrl = trainer.ProfileImageUrl,
+                    TrainerName = trainer.FullName
+                };
+
+                myCoursesViewModel.Add(viewModel);
+            }
 
             return View(myCoursesViewModel);
         }
@@ -166,12 +220,12 @@
 
         [Authorize]
         public async Task<IActionResult> Details(int id)
-        {         
+        {
             var courseDetails = await courseService.GetCourseDetails(id);
-                 
+
             var courseDetailsViewModel = new CourseDetailsViewModel
             {
-                CourseServiceModel = courseDetails,              
+                CourseServiceModel = courseDetails,
             };
 
             return View(courseDetailsViewModel);
@@ -181,6 +235,14 @@
         [Authorize]
         public async Task<IActionResult> DeleteCourse(int id)
         {
+            int courseCreatorId = await this.courseService.GetCourseCreatorId(id);
+            int currentTrainerId = await this.trainerService.GetTrainerIdByUserId(User.Id());
+
+            if (currentTrainerId != courseCreatorId)
+            {
+                return Unauthorized();
+            }
+
             var result = await courseService.DeleteCourse(id);
 
             if (result == false)
